refactor(spectral tools): ispectrogrampainter is internal
