+++ src/spectrogram/CMakeLists.txt
@@ -8,12 +8,12 @@ set(MODULE_SRC
     ${CMAKE_CURRENT_LIST_DIR}/ispectrogramconfiguration.h
     ${CMAKE_CURRENT_LIST_DIR}/iglobalspectrogramconfiguration.h
     ${CMAKE_CURRENT_LIST_DIR}/itrackspectrogramconfiguration.h
-    ${CMAKE_CURRENT_LIST_DIR}/ispectrogrampainter.h
     ${CMAKE_CURRENT_LIST_DIR}/ispectrogramservice.h
     ${CMAKE_CURRENT_LIST_DIR}/spectrogrammodule.cpp
     ${CMAKE_CURRENT_LIST_DIR}/spectrogrammodule.h
     ${CMAKE_CURRENT_LIST_DIR}/spectrogramtypes.h
 
+    ${CMAKE_CURRENT_LIST_DIR}/internal/ispectrogrampainter.h
     ${CMAKE_CURRENT_LIST_DIR}/internal/iclipchannelreader.h
     ${CMAKE_CURRENT_LIST_DIR}/internal/clipparameters.cpp
     ${CMAKE_CURRENT_LIST_DIR}/internal/clipparameters.h
