+++ src/spectrogram/internal/ispectrogrampainter.h
